move generic list into component directory
