+++ src/app/[locale]/list/_components/genericList/GenericAasList.tsx
@@ -1,5 +1,5 @@
 ﻿import { Table, TableBody, TableCell, TableContainer, TableHead, TableRow } from '@mui/material';
-import { GenericAasListEntry } from 'app/[locale]/viewer/discovery/_components/GenericAasListEntry';
+import { GenericAasListEntry } from 'app/[locale]/list/_components/genericList/GenericAasListEntry';
 import { AasListConfig, AasListEntry } from 'lib/types/AasListEntry';
 import { useTranslations } from 'next-intl';
