+++ src/app/[locale]/list/_components/genericList/GenericAasListEntry.tsx
