+++ src/app/[locale]/viewer/discovery/_components/DiscoveryListView.tsx
@@ -9,7 +9,7 @@ import { performSearchAasFromAllRepositories } from 'lib/services/repository-acc
 import { useTranslations } from 'next-intl';
 import { LocalizedError } from 'lib/util/LocalizedError';
 import { AasListEntry } from 'lib/types/AasListEntry';
-import { GenericListDataWrapper } from 'app/[locale]/viewer/discovery/_components/GenericListDataWrapper';
+import { GenericListDataWrapper } from 'app/[locale]/list/_components/genericList/GenericListDataWrapper';
 
 async function getRepositoryUrl(aasId: string): Promise<string | undefined> {
     const registrySearchResult = await performRegistryAasSearch(aasId);
